refactor: clean up code and resolve warnings
