+++ shell/crates/app_drawer/src/models.rs
@@ -1,12 +1,9 @@
-use crate::prelude::IconName;
-use crate::ui::utils::prelude::{DesktopApp, DesktopApps};
+use crate::ui::utils::prelude::DesktopApps;
 use gpui::*;
 
 #[derive(Default, Clone)]
 pub struct AppDrawerState {
-    pub search: String,
     pub apps: DesktopApps,
-    pub recent: Vec<String>,
 }
 
 impl Global for AppDrawerState {}
