+++ shell/crates/app_drawer/src/ui/utils/desktop_apps.rs
@@ -1,8 +1,5 @@
-use std::collections::BTreeMap;
 use std::collections::HashMap;
-use std::ffi::OsString;
-use std::path::{Path, PathBuf};
-use std::process::Command;
+use std::path::{PathBuf};
 
 use crate::prelude::Icon;
 use crate::prelude::IconName;
