+++ shell/crates/app_drawer/src/ui/widgets/mod.rs
@@ -5,3 +5,6 @@ mod text_button;
 pub use icon_button::*;
 pub use sub_window::*;
 pub use text_button::*;
+
+mod bottom_sheet;
+pub use bottom_sheet::*;
